Do not emit cross-platform artifacts. Instead, always expand the individual configs completely.
If we try to determine the set of "common" artifacts by excluding the current machine config, they don't get resolved correctly if the machine specific config contains newer versions of a given dependency.

Instead, we simply include all the artifacts on the machine specific configs, which will get resolved correctly.