Conveyor: go back to focal as the base distro for debs.
This fixes a regression in which JavaFX apps could no longer be installed on Focal due to the Debian guys renaming the gdk-pixbuf package. There's a transitional package in Jammy so old packages can still install on newer releases, but this doesn't work if we target a newer release by default. Focal came out in 2020 so it's already kind of an aggressive default.